how to play 재시작기능 추가
